Fix immutable vs locked issue
