+++ features/folders/update-folder/index.ts
@@ -1,5 +1,5 @@
 export { updateFolderAction } from "./update-folder.action";
 export type { UpdateFolderResult } from "./update-folder.action";
-export { FolderEditDialog } from "./ui/folder-edit-dialog";
+export { FolderEditDialog } from "./ui/edit-folder-dialog";
 export { FolderSettingsCard } from "@/features/folders/ui/folder-settings-card";
 export type { FolderSettingsCardProps } from "@/features/folders/ui/folder-settings-card";
